Update Home Component to dispatch an action to reducer
On the Home component: when we get the values from the API, I want to dispatch an action, which goes to the reducer, the reducer then updates the state.

Notice code reordering as: React Hook "useDispatch" cannot be called inside a callback. React Hooks must be called in a React function component or a custom React Hook function.